requirements.txt
langchain
langchain-groq
langgraph
qdrant-client
python-dotenv
requests
pypdf
streamlit
pytest
langchain_community
